# backend/api-spec.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Artifex Gallery API",
    "version": "1.0.0",
    "description": "AI-powered image and video gallery with auto-tagging, NSFW detection, and federation support."
  },
  "servers": [{ "url": "/api" }],
  "components": {
    "securitySchemes": {
      "bearerAuth": { "type": "http", "scheme": "bearer", "bearerFormat": "JWT" }
    },
...
  },
  "paths": {
...
    "/tags": { "get": { "tags": ["Tags"], "summary": "List all tags with counts", "parameters": [{ "name": "category", "in": "query", "schema": { "type": "string" } }], "responses": { "200": { "description": "Tag list" } } } },
...
    "/tags/search": { "get": { "tags": ["Tags"], "summary": "Search/autocomplete tags", "parameters": [{ "name": "q", "in": "query", "schema": { "type": "string" } }], "responses": { "200": { "description": "Matching tags" } } } },
    "/tags/image/{imageId}": {
      "get": { "tags": ["Tags"], "summary": "Get tags for image", "parameters": [{ "name": "imageId", "in": "path", "required": true, "schema": { "type": "integer" } }], "responses": { "200": { "description": "Tags" } } },
      "post": { "tags": ["Tags"], "summary": "Add manual tag", "security": [{ "bearerAuth": [] }], "parameters": [{ "name": "imageId", "in": "path", "required": true, "schema": { "type": "integer" } }], "responses": { "200": { "description": "Tag added" } } }
    },
    "/tags/image/{imageId}/{tagId}": {
      "delete": { "tags": ["Tags"], "summary": "Remove tag from image", "security": [{ "bearerAuth": [] }], "parameters": [{ "name": "imageId", "in": "path", "required": true, "schema": { "type": "integer" } }, { "name": "tagId", "in": "path", "required": true, "schema": { "type": "integer" } }], "responses": { "200": { "description": "Removed" } } }
    },
    "/tags/vision/batch": { "post": { "tags": ["ML"], "summary": "Batch vision tagging", "security": [{ "bearerAuth": [] }], "responses": { "200": { "description": "Processed count" } } } },
    "/tags/nsfw/batch": { "post": { "tags": ["ML"], "summary": "Batch NSFW detection", "security": [{ "bearerAuth": [] }], "responses": { "200": { "description": "Processed count" } } } },
    "/tags/caption/batch": { "post": { "tags": ["ML"], "summary": "Batch captioning", "security": [{ "bearerAuth": [] }], "responses": { "200": { "description": "Processed count" } } } },
    "/tags/backfill": { "post": { "tags": ["ML"], "summary": "Backfill metadata tags", "security": [{ "bearerAuth": [] }], "responses": { "200": { "description": "Tagged count" } } } },
    "/tags/jobs/stats": { "get": { "tags": ["ML"], "summary": "Job queue statistics", "security": [{ "bearerAuth": [] }], "responses": { "200": { "description": "Queue stats" } } } },
    "/tags/jobs/retry": { "post": { "tags": ["ML"], "summary": "Retry all failed jobs", "security": [{ "bearerAuth": [] }], "responses": { "200": { "description": "Retried count" } } } },
    "/tags/jobs/cleanup": { "delete": { "tags": ["ML"], "summary": "Clean completed jobs", "security": [{ "bearerAuth": [] }], "responses": { "200": { "description": "Cleaned count" } } } },
...
  }
}